+++ xotl/crdt/base.py
@@ -11,11 +11,11 @@ from dataclasses import dataclass
 
 
 @dataclass(frozen=True, order=True)
-class Actor:
-    '''Represents an actor or node that holds replicated objects.
+class Process:
+    '''Represents a process or node that holds replicated objects.
 
-    We require (for some CRDTs) that actors are uniquely named and totally
-    ordered across the cluster.  So when adding/removing actors you should
+    We require (for some CRDTs) that processes are uniquely named and totally
+    ordered across the cluster.  So when adding/removing a process you should
     take measures for not reusing old names.
 
     '''
@@ -27,10 +27,10 @@ class Actor:
         object.__setattr__(self, 'order', order)
 
     def __repr__(self):
-        return f"Actor({self.name!r}, {self.order!r})"
+        return f"Process({self.name!r}, {self.order!r})"
 
     def __eq__(self, other) -> bool:
-        if isinstance(other, Actor):
+        if isinstance(other, Process):
             return self.name == other.name
         else:
             return NotImplemented
@@ -43,8 +43,8 @@ class CvRDT:
     **must** implement the following methods and attributes.
 
     '''
-    def __init__(self, *, actor: Actor) -> None:
-        self.actor = actor
+    def __init__(self, *, process: Process) -> None:
+        self.process = process
         self.init()
 
     def init(self) -> None: