+++ xotl/crdt/counter.py
@@ -18,11 +18,11 @@ class GCounter(CvRDT):
         self.vclock = VClock()
 
     def __repr__(self):
-        return f"<GCounter of {self.value}; {self.actor}, {self.vclock.simplified}>"
+        return f"<GCounter of {self.value}; {self.process}, {self.vclock.simplified}>"
 
     def incr(self):
         'Increases the counter by one.'
-        self.vclock = self.vclock.bump(self.actor)
+        self.vclock = self.vclock.bump(self.process)
 
     @property
     def value(self) -> int:
@@ -43,14 +43,14 @@ class GCounter(CvRDT):
         '''Reset the counter to 0.
 
         .. warning:: This an operation that must be coordinated between
-           actors.
+           processes.
 
         '''
         self.vclock.reset()
 
     def __eq__(self, other) -> bool:
         if isinstance(other, GCounter):
-            return self.actor == other.actor and self.vclock == other.vclock
+            return self.process == other.process and self.vclock == other.vclock
         else:
             return NotImplemented
 
@@ -58,8 +58,8 @@ class GCounter(CvRDT):
 class PNCounter(CvRDT):
     '''A counter that allows increments and decrements.'''
     def init(self):
-        self.pos = GCounter(actor=self.actor)
-        self.neg = GCounter(actor=self.actor)
+        self.pos = GCounter(process=self.process)
+        self.neg = GCounter(process=self.process)
 
     def __repr__(self):
         return f"<PNCounter of {self.value}; with {self.pos} and {self.neg}>"
@@ -90,7 +90,7 @@ class PNCounter(CvRDT):
 
     def __eq__(self, other):
         if isinstance(other, PNCounter):
-            return (self.actor == other.actor and
+            return (self.process == other.process and
                     self.pos == other.pos and
                     self.neg == other.neg)
         else:
@@ -100,7 +100,7 @@ class PNCounter(CvRDT):
         '''Reset the counter to 0.
 
         .. warning:: This an operation that must be coordinated between
-           actors.
+           processes.
 
         '''
         self.pos.reset()