+++ docs/source/history.rst
@@ -8,12 +8,14 @@ Series 0.x
 Unreleased.  Release 0.2.0
 --------------------------
 
+- Replace 'process' for 'actor' across the code-base.
+
 - Extract the state dump/reconstruction from `~xotl.crdt.base.CvRDT`:class:;
   add functions `~xotl.crdt.base.get_state`:func: and
   `~xotl.crdt.base.from_state`:func:.
 
-- Add `~xotl.crdt.base.Actor`:class: to capture the required interface of
-  actors.
+- Add `~xotl.crdt.base.Process`:class: to capture the required interface of
+  processes.
 
 
 2018-10-03.  Release 0.1.1